Rename Frosting recipe to Cake.Frosting.Issues.Recipe
