+++ .github/dependabot.yml
@@ -2,7 +2,7 @@
 version: 2
 updates:
 - package-ecosystem: nuget
-  directory: "/Cake.Issues.FrostingRecipe"
+  directory: "/Cake.Frosting.Issues.Recipe"
   schedule:
     interval: daily
   open-pull-requests-limit: 10
