+++ Cake.Frosting.Issues.Recipe/Cake.Frosting.Issues.Recipe.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30804.86
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cake.Frosting.Issues.Recipe", "Cake.Frosting.Issues.Recipe\Cake.Frosting.Issues.Recipe.csproj", "{EF4F2348-04B1-4BDD-87A3-1622845227A9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Debug|x64.Build.0 = Debug|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Debug|x86.Build.0 = Debug|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Release|x64.ActiveCfg = Release|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Release|x64.Build.0 = Release|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Release|x86.ActiveCfg = Release|Any CPU
+		{EF4F2348-04B1-4BDD-87A3-1622845227A9}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {DC4F437B-C35D-481A-B0E2-8549E2BDEB8D}
+	EndGlobalSection
+EndGlobal
