+++ Cake.Frosting.Issues.Recipe/Cake.Frosting.Issues.Recipe/PullRequestSystems/GitHubPullRequestSystem.cs
@@ -1,0 +1,39 @@
+using Cake.Issues;
+
+namespace Cake.Frosting.Issues.Recipe
+{
+    /// <summary>
+    /// Support for GitHub hosted code.
+    /// </summary>
+    internal class GitHubPullRequestSystem : BasePullRequestSystem
+    {
+        /// <inheritdoc />
+        public override void ReportIssuesToPullRequest(IssuesContext context)
+        {
+            context.NotNull(nameof(context));
+
+            // Not supported yet
+        }
+
+        /// <inheritdoc />
+        public override void SetPullRequestIssuesState(IssuesContext context)
+        {
+            context.NotNull(nameof(context));
+
+            // Not supported yet
+        }
+
+        /// <inheritdoc />
+        public override FileLinkSettings GetFileLinkSettings(IssuesContext context)
+        {
+            context.NotNull(nameof(context));
+
+            var rootPath = context.State.RepositoryRootDirectory.GetRelativePath(context.State.BuildRootDirectory);
+
+            return context.IssueFileLinkSettingsForGitHubCommit(
+                context.State.RepositoryRemoteUrl,
+                context.State.CommitId,
+                rootPath.FullPath);
+        }
+    }
+}
