--- Cake.Issues.FrostingRecipe/Cake.Issues.FrostingRecipe/IssuesTask.cs
@@ -1,17 +1,0 @@
-﻿using Cake.Frosting;
-
-namespace Cake.Issues.FrostingRecipe
-{
-    /// <summary>
-    /// Main task for issue management integration.
-    /// </summary>
-    [TaskName("Issues")]
-    [IsDependentOn(typeof(PublishIssuesArtifactsTask))]
-    [IsDependentOn(typeof(ReportIssuesToBuildServerTask))]
-    [IsDependentOn(typeof(CreateSummaryIssuesReportTask))]
-    [IsDependentOn(typeof(ReportIssuesToPullRequestTask))]
-    [IsDependentOn(typeof(SetPullRequestIssuesStateTask))]
-    public sealed class IssuesTask : FrostingTask<IssuesContext>
-    {
-    }
-}
