--- Cake.Issues.FrostingRecipe/Cake.Issues.FrostingRecipe/PublishIssuesArtifactsTask.cs
@@ -1,32 +1,0 @@
-﻿using Cake.Common.Build;
-using Cake.Common.Diagnostics;
-using Cake.Frosting;
-
-namespace Cake.Issues.FrostingRecipe
-{
-    /// <summary>
-    /// Publish issue artifacts to build server.
-    /// </summary>
-    [TaskName("Publish-IssuesArtifacts")]
-    [IsDependentOn(typeof(CreateFullIssuesReportTask))]
-    public sealed class PublishIssuesArtifactsTask : FrostingTask<IssuesContext>
-    {
-        /// <inheritdoc/>
-        public override bool ShouldRun(IssuesContext context)
-        {
-            return !context.BuildSystem().IsLocalBuild;
-        }
-
-        /// <inheritdoc/>
-        public override void Run(IssuesContext context)
-        {
-            if (context.State.BuildServer == null)
-            {
-                context.Information("Not supported build server.");
-                return;
-            }
-
-            context.State.BuildServer.PublishIssuesArtifacts(context);
-        }
-    }
-}
