--- Cake.Issues.FrostingRecipe/Cake.Issues.FrostingRecipe/PullRequestSystems/BasePullRequestSystem.cs
@@ -1,20 +1,0 @@
-namespace Cake.Issues.FrostingRecipe
-{
-    /// <summary>
-    /// Basic implementation for all pull request server.
-    /// </summary>
-    internal abstract class BasePullRequestSystem : IIssuesPullRequestSystem
-    {
-        /// <inheritdoc />
-        public abstract void ReportIssuesToPullRequest(
-            IssuesContext context);
-
-        /// <inheritdoc />
-        public abstract void SetPullRequestIssuesState(
-            IssuesContext context);
-
-        /// <inheritdoc />
-        public abstract FileLinkSettings GetFileLinkSettings(
-            IssuesContext context);
-    }
-}
